% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_srv.R
\name{overview_srv}
\alias{overview_srv}
\title{Server Logic for Overview Module}
\usage{
overview_srv(id, r, verbose = FALSE)
}
\arguments{
\item{id}{Module ID.}

\item{r}{Reactive values object for shared data.}

\item{verbose}{Logical, whether to print verbose output. Defaults to FALSE.}
}
\value{
None. Handles server-side logic for the overview module.
}
\description{
Server Logic for Overview Module
}
