% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_ui.R
\name{overview_ui}
\alias{overview_ui}
\title{Overview UI Module}
\usage{
overview_ui(id)
}
\arguments{
\item{id}{The unique identifier for the module namespace.}
}
\value{
A Shiny UI element.
}
\description{
This function generates the user interface for the overview module in a Shiny application.
}
