% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfolioselector_ui.R
\name{pfolioselector_ui}
\alias{pfolioselector_ui}
\title{Portfolio Selector UI}
\usage{
pfolioselector_ui(id, con)
}
\arguments{
\item{id}{Character string specifying the module's namespace ID.}

\item{con}{Database connection object used for retrieving portfolio data.}
}
\value{
A Shiny UI component for portfolio selection.
}
\description{
Creates the user interface for a Shiny module that allows users to select a portfolio.
}
