% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebal_ui.R
\name{rebal_ui}
\alias{rebal_ui}
\title{Rebalancing UI Module}
\usage{
rebal_ui(id)
}
\arguments{
\item{id}{Character string specifying the module's namespace ID.}
}
\value{
A Shiny UI component for the rebal module.
}
\description{
This function generates the user interface for the rebalancing Shiny module.
}
