% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bfett.R
\name{test_bfett}
\alias{test_bfett}
\title{Runs tests}
\usage{
test_bfett(fn = NULL, which = c("summary", "fails", "all"), verbose = FALSE)
}
\arguments{
\item{fn}{file name(s) of the tests to run. Default (NULL) runs all tests.}

\item{which}{character, either "summary" (default, aggregates the results),
"fails" (returns tests that failed), or "all"}

\item{verbose}{FALSE (default) silent, TRUE diagnostic messages}
}
\value{
a data.frame
}
\description{
Runs tests
}
