% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_srv.R
\name{treemap_srv}
\alias{treemap_srv}
\title{Treemap Server Module}
\usage{
treemap_srv(id, r, verbose = FALSE)
}
\arguments{
\item{id}{Character string specifying the module's namespace ID.}

\item{r}{Reactive values object for sharing data within the module.}

\item{verbose}{Logical indicating whether to enable verbose output. Defaults to FALSE.}
}
\value{
None. Handles server-side logic for the treemap module.
}
\description{
Server logic for the treemap Shiny module.
}
