% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_ui.R
\name{treemap_ui}
\alias{treemap_ui}
\title{Treemap UI Module}
\usage{
treemap_ui(id, con)
}
\arguments{
\item{id}{A character string specifying the module's namespace ID.}

\item{con}{A database connection object used for data retrieval.}
}
\value{
A Shiny UI element for rendering the treemap.
}
\description{
This function generates the user interface for a treemap module in a Shiny application.
}
