% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_irr.R
\name{val_irr}
\alias{val_irr}
\title{Calculate the Internal Rate of Return (IRR) for a portfolio}
\usage{
val_irr(con, portfolio, trans = NULL, current_value = NULL)
}
\arguments{
\item{con}{Database connection object.}

\item{portfolio}{Character string specifying the portfolio identifier.}

\item{trans}{either NULL (default) or data.frame containing transaction data. If NULL, reads from the database.}

\item{current_value}{either NULL (default) or data.frame containing the current value. If NULL, reads from the database.}
}
\value{
A numeric value representing the annualized IRR or NA if something went wrong.
}
\description{
This function computes the Internal Rate of Return (IRR) based on cash flows
from a portfolio. Handles deposits and withdrawals.
}
\details{
It calculates numerically a daily rate and returns 360 times the rate.
}
