% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_pfoliovalue.R
\name{val_pfoliovalue}
\alias{val_pfoliovalue}
\title{Calculate the value of a portfolio}
\usage{
val_pfoliovalue(con, portfolio)
}
\arguments{
\item{con}{Database connection object.}

\item{portfolio}{Character string specifying the portfolio identifier.}
}
\value{
data.frame with columns calendar_week, typ ("activepos", "cash"), and value
}
\description{
This function computes the total value of a given portfolio based on the provided database connection.
}
